refact: move syncIndexToMarketPrice, syncMarkPriceToMarketPrice to test/shared/utilities.ts
