+++ test/shared/utilities.ts
@@ -1,11 +1,11 @@
 import { BaseContract, BigNumber, BigNumberish } from "ethers"
-import { BaseToken, Exchange, UniswapV3Pool, VirtualToken } from "../../typechain"
+import { BaseToken, Exchange, TestAccountBalance, UniswapV3Pool, VirtualToken } from "../../typechain"
 
-import { LogDescription } from "@ethersproject/abi"
 import { MockContract } from "@eth-optimism/smock"
+import { LogDescription } from "@ethersproject/abi"
 import { TransactionReceipt } from "@ethersproject/abstract-provider"
 import bn from "bignumber.js"
-import { parseUnits } from "ethers/lib/utils"
+import { parseEther, parseUnits } from "ethers/lib/utils"
 
 bn.config({ EXPONENTIAL_AT: 999999, DECIMAL_PLACES: 40 })
 
@@ -79,6 +79,17 @@ export async function syncIndexToMarketPrice(aggregator: MockContract, pool: Uni
     aggregator.smocked.latestRoundData.will.return.with(async () => {
         return [0, parseUnits(price, oracleDecimals), 0, 0, 0]
     })
+}
+
+export async function syncMarkPriceToMarketPrice(
+    accountBalance: TestAccountBalance,
+    baseToken: string,
+    pool: UniswapV3Pool,
+) {
+    const slot0 = await pool.slot0()
+    const sqrtPrice = slot0.sqrtPriceX96
+    const price = formatSqrtPriceX96ToPrice(sqrtPrice, 18)
+    await accountBalance.mockMarkPrice(baseToken, parseEther(price))
 }
 
 export async function getMarketTwap(exchange: Exchange, baseToken: BaseToken, interval: number) {
